+++ lib/measures/generic_qaqc/resources/check_plant_temps.rb
@@ -43,8 +43,8 @@ module OsLib_QAQC
           when 'OS_SetpointManager_Scheduled'
             sch = spm.to_SetpointManagerScheduled.get.schedule
             if sch.to_ScheduleRuleset.is_initialized
-              min_c = std.schedule_ruleset_annual_min_max_value(sch.to_ScheduleRuleset.get)['min']
-              max_c = std.schedule_ruleset_annual_min_max_value(sch.to_ScheduleRuleset.get)['max']
+              min_c = openstudiostandards::schedules.schedule_ruleset_annual_min_max_value(sch.to_ScheduleRuleset.get)['min']
+              max_c = openstudiostandards::schedules.schedule_ruleset_annual_min_max_value(sch.to_ScheduleRuleset.get)['max']
             elsif sch.to_ScheduleConstant.is_initialized
               min_c = std.schedule_constant_annual_min_max_value(sch.to_ScheduleConstant.get)['min']
               max_c = std.schedule_constant_annual_min_max_value(sch.to_ScheduleConstant.get)['max']
@@ -60,8 +60,8 @@ module OsLib_QAQC
             next if low_sch.empty?
             low_sch = low_sch.get
             if low_sch.to_ScheduleRuleset.is_initialized
-              min_c = std.schedule_ruleset_annual_min_max_value(low_sch.to_ScheduleRuleset.get)['min']
-              max_c = std.schedule_ruleset_annual_min_max_value(low_sch.to_ScheduleRuleset.get)['max']
+              min_c = openstudiostandards::schedules.schedule_ruleset_annual_min_max_value(low_sch.to_ScheduleRuleset.get)['min']
+              max_c = openstudiostandards::schedules.schedule_ruleset_annual_min_max_value(low_sch.to_ScheduleRuleset.get)['max']
             elsif low_sch.to_ScheduleConstant.is_initialized
               min_c = std.schedule_constant_annual_min_max_value(low_sch.to_ScheduleConstant.get)['min']
               max_c = std.schedule_constant_annual_min_max_value(low_sch.to_ScheduleConstant.get)['max']
@@ -74,8 +74,8 @@ module OsLib_QAQC
             next if high_sch.empty?
             high_sch = high_sch.get
             if high_sch.to_ScheduleRuleset.is_initialized
-              min_c = std.schedule_ruleset_annual_min_max_value(high_sch.to_ScheduleRuleset.get)['min']
-              max_c = std.schedule_ruleset_annual_min_max_value(high_sch.to_ScheduleRuleset.get)['max']
+              min_c = openstudiostandards::schedules.schedule_ruleset_annual_min_max_value(high_sch.to_ScheduleRuleset.get)['min']
+              max_c = openstudiostandards::schedules.schedule_ruleset_annual_min_max_value(high_sch.to_ScheduleRuleset.get)['max']
             elsif high_sch.to_ScheduleConstant.is_initialized
               min_c = std.schedule_constant_annual_min_max_value(high_sch.to_ScheduleConstant.get)['min']
               max_c = std.schedule_constant_annual_min_max_value(high_sch.to_ScheduleConstant.get)['max']